# openapi.yaml
openapi: 3.0.3
info:
  title: NexusID API
  version: 0.1.0
  description: OAuth2/OIDC, admin APIs, and user portal APIs.
servers:
  - url: http://127.0.0.1:4000
paths:
...
  /api/admin/saml/service-providers:
    get:
      summary: List SAML service providers
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: enabled
          schema:
            type: string
            enum: ["true", "false"]
      responses:
        '200':
          description: List of SAML service providers
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/SamlServiceProvider'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
    post:
      summary: Create SAML service provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [entityId, acsUrl, signingCertificate]
              properties:
                entityId:
                  type: string
                  description: SAML service provider entity ID
                acsUrl:
                  type: string
                  format: uri
                  description: Assertion Consumer Service URL
                sloUrl:
                  type: string
                  format: uri
                  description: Single Logout Service URL (optional)
                signingCertificate:
                  type: string
                  description: X.509 certificate for signature verification
                encryptionCertificate:
                  type: string
                  description: X.509 certificate for response encryption (optional)
                nameIdFormat:
                  type: string
                  enum: [persistent, transient, emailAddress]
                  default: persistent
                enabled:
                  type: boolean
                  default: true
      responses:
        '201':
          description: SAML service provider created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SamlServiceProvider'
  /api/admin/saml/service-providers/{id}:
    get:
      summary: Get SAML service provider by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: SAML service provider details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SamlServiceProvider'
        '404':
          description: Service provider not found
    patch:
      summary: Update SAML service provider
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                entityId:
                  type: string
                acsUrl:
                  type: string
                  format: uri
                sloUrl:
                  type: string
                  format: uri
                signingCertificate:
                  type: string
                encryptionCertificate:
                  type: string
                nameIdFormat:
                  type: string
                  enum: [persistent, transient, emailAddress]
                enabled:
                  type: boolean
      responses:
        '200':
          description: SAML service provider updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SamlServiceProvider'
    delete:
      summary: Delete SAML service provider
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: SAML service provider deleted
  /api/admin/saml/service-providers/{id}/metadata:
    post:
      summary: Upload and parse service provider metadata XML
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [metadata]
              properties:
                metadata:
                  type: string
                  description: Raw SAML metadata XML document.
                overwriteManualFields:
                  type: boolean
                  default: true
                  description: When true, parsed metadata values overwrite entityId/ACS/SLO/certificate fields.
      responses:
        '200':
          description: Metadata uploaded and parsed
          content:
            application/json:
              schema:
                type: object
                properties:
                  serviceProvider:
                    $ref: '#/components/schemas/SamlServiceProvider'
                  imported:
                    type: object
                    properties:
                      entityId:
                        type: string
                      acsUrl:
                        type: string
                      sloUrl:
                        type: string
                      hasSigningCertificate:
                        type: boolean
        '400':
          description: Invalid metadata payload or service provider id
  /api/admin/saml/service-providers/{id}/certificates/rotate:
    post:
      summary: Rotate service provider signing or encryption certificate
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [certificate]
              properties:
                certificateType:
                  type: string
                  enum: [signing, encryption]
                  default: signing
                certificate:
                  type: string
      responses:
        '200':
          description: Certificate rotated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SamlServiceProvider'
        '400':
          description: Invalid rotation payload or service provider id
  /api/admin/saml/assertions:
    get:
      summary: List SAML assertion audits
      parameters:
        - in: query
          name: spId
          schema:
            type: string
          description: Filter by service provider ID
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: startDate
          schema:
            type: string
            format: date-time
          description: Filter assertions from this date onwards
        - in: query
          name: endDate
          schema:
            type: string
            format: date-time
          description: Filter assertions up to this date
      responses:
        '200':
          description: List of SAML assertion audits
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/SamlAssertionAudit'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
...
components:
  schemas:
...
    SamlServiceProvider:
      type: object
      properties:
        id:
          type: string
          description: Service provider ID
        entityId:
          type: string
          description: SAML entity ID
        acsUrl:
          type: string
          format: uri
          description: Assertion Consumer Service URL
        sloUrl:
          type: string
          format: uri
          nullable: true
          description: Single Logout Service URL
        signingCertificate:
          type: string
          description: X.509 certificate for signature verification
        encryptionCertificate:
          type: string
          nullable: true
          description: X.509 certificate for response encryption
        nameIdFormat:
          type: string
          enum: [persistent, transient, emailAddress]
          description: NameID format for assertions
        enabled:
          type: boolean
          description: Whether the service provider is enabled
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SamlAssertionAudit:
      type: object
      properties:
        id:
          type: string
          description: Audit record ID
        spId:
          type: string
          description: Service provider ID that processed the assertion
        requestId:
          type: string
          description: Upstream request ID correlated to this response
        responseId:
          type: string
          description: SAML response ID
        subject:
          type: string
          description: Assertion subject/NameID
        audience:
          type: string
          description: Intended assertion audience
        assertionId:
          type: string
          description: SAML assertion ID
        issueInstant:
          type: string
          format: date-time
        notOnOrAfter:
          type: string
          format: date-time
        destinationUrl:
          type: string
          format: uri
        statusCode:
          type: string
          description: SAML status code URN from assertion processing
        createdAt:
          type: string
          format: date-time
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT